重构梯度计算并移除 FP8 存储压缩支持
- 重构梯度计算逻辑，在 Python 回调中使用 PyTorch 的 `torch.autograd.backward` (VJP) 来计算物理参数梯度，取代此前原本在 C++/CUDA/Metal 中手写的 kernel。
- 从各个后端的 C++ 源码（如 `maxwell.cu`, `maxwell_metal.mm` 和 `maxwell.c`）中移除 `convert_grad_ca_cb_to_eps_sigma` 及其相关逻辑。
- 移除不再需要的 `autograd_utils.py`。
- 移除 FP8 存储压缩相关支持，包括清理 `storage.py` 及 C++ 后尾相关逻辑，并删除 `test_fp8_storage.py` 测试。
- 更新 `uv.lock` 中 `tide-gpr` 版本号至 0.0.20。

+++ examples/example_multiscale_crosscorr_wrong_wavelet.py
@@ -25,7 +25,7 @@ first_source = 0
 batch_size = 16
 model_gradient_sampling_interval = 5
 storage_mode = "device"
-storage_compression = "fp8"
+storage_compression = "bf16"
 
 model_path = "examples/data/OverThrust.npy"
 epsilon_true_raw = np.load(model_path)
